@@ -1,7 +1,7 @@
 ﻿<Project Sdk="Microsoft.NET.Sdk.Web">
 
   <PropertyGroup>
-    <TargetFramework>netcoreapp2.2</TargetFramework>
+    <TargetFramework>netcoreapp3.0</TargetFramework>
     <AssetTargetFallback>$(AssetTargetFallback);portable-net45+win8+wp8+wpa81;</AssetTargetFallback>
     <PreserveCompilationContext>true</PreserveCompilationContext>
     <AssemblyName>FacadeCompanyName.FacadeProjectName.Web.Host</AssemblyName>
@@ -21,14 +21,13 @@
   </ItemGroup>
 
   <ItemGroup>
-    <PackageReference Include="Facade.AspNetCore" Version="1.2.6" />
-    <PackageReference Include="Facade.AspNetCore.Zero" Version="1.2.6" />
-    <PackageReference Include="Facade.Castle.NLogger" Version="1.2.6" />
-    <PackageReference Include="Facade.AspNetCore.SignalR" Version="1.2.6" />
-    <PackageReference Include="Microsoft.AspNetCore.App" />
-    <PackageReference Include="Microsoft.AspNetCore.Razor.Design" Version="2.2.0" PrivateAssets="All" />
+    <PackageReference Include="Facade.AspNetCore" Version="2.0.0" />
+    <PackageReference Include="Facade.AspNetCore.Zero" Version="2.0.0" />
+    <PackageReference Include="Facade.Castle.NLogger" Version="2.0.0" />
+    <PackageReference Include="Facade.AspNetCore.SignalR" Version="2.0.0" />  
 
-    <PackageReference Include="Swashbuckle.AspNetCore" Version="4.0.1" />
+    <PackageReference Include="Microsoft.AspNetCore.Authentication.JwtBearer" Version="3.0.0" />
+    <PackageReference Include="Swashbuckle.AspNetCore" Version="5.0.0-rc4" />
   </ItemGroup>
 
   <ItemGroup>
