@@ -1,5 +1,7 @@
 ﻿using FacadeCompanyName.FacadeProjectName.DomainService.Demo;
 using FacadeCompanyName.FacadeProjectName.DomainService.Demo.Dto;
+using FacadeCompanyName.FacadeProjectName.DomainService.Share;
+using System;
 using System.Threading.Tasks;
 
 namespace FacadeCompanyName.FacadeProjectName.Application.Demo
@@ -7,9 +9,11 @@ namespace FacadeCompanyName.FacadeProjectName.Application.Demo
     public class DemoApplication : FacadeProjectNameApplicationBase, IDemoApplication
     {
         private readonly IDemoService _demoService;
-        public DemoApplication(IDemoService demoService)
+        private readonly IAppQueryRepository _appQueryRepository;
+        public DemoApplication(IDemoService demoService, IAppQueryRepository appQueryRepository)
         {
             _demoService = demoService;
+            _appQueryRepository = appQueryRepository;
         }
 
         public async Task<string> Check(CheckInput input)
@@ -19,6 +23,10 @@ namespace FacadeCompanyName.FacadeProjectName.Application.Demo
         public async Task<string> Query(int id)
         {
             return await _demoService.Query(id);
+        }
+        public async Task Health()
+        {
+            await _appQueryRepository.QueryAsync<DateTime>("select sysdate from dual");
         }
     }
 }
