@@ -24,7 +24,7 @@ namespace FacadeCompanyName.FacadeProjectName.Application
             IocManager.RegisterAssemblyByConvention(thisAssembly);
             Configuration.Modules.AbpAutoMapper().Configurators.Add(
                 // Scan the assembly for classes which inherit from AutoMapper.Profile
-                cfg => cfg.AddProfiles(thisAssembly)
+                cfg => cfg.AddMaps(thisAssembly)
             );
         }
 
