@@ -1,4 +1,5 @@
-﻿using Facade.AspNetCore.Web.Models;
+﻿using Facade.AspNetCore.Mvc.Authorization;
+using Facade.AspNetCore.Web.Models;
 using FacadeCompanyName.FacadeProjectName.Application.Demo;
 using FacadeCompanyName.FacadeProjectName.DomainService.Demo.Dto;
 using Microsoft.AspNetCore.Mvc;
@@ -14,6 +15,16 @@ namespace FacadeCompanyName.FacadeProjectName.Web.Host.Controllers
         {
             _demoApplication = demoApplication;
         }
+
+        [Route("health")]
+        [HttpGet]
+        [NoToken]
+        public async Task<JsonResponse> Health()
+        {
+            await _demoApplication.Health();
+            return await Task.FromResult(new JsonResponse(true, L("WelcomeMessage")));
+        }
+
         [Route("check")]
         [HttpPost]
         public async Task<JsonResponse<string>> Check([FromBody]CheckInput input)
