@@ -7,6 +7,7 @@ using Microsoft.AspNetCore;
 using Microsoft.AspNetCore.Hosting;
 using Microsoft.Extensions.Configuration;
 using Microsoft.Extensions.Logging;
+using Facade.AspNetCore.Web.NLog;
 
 namespace FacadeCompanyName.FacadeProjectName.Web.Host
 {
@@ -18,7 +19,12 @@ namespace FacadeCompanyName.FacadeProjectName.Web.Host
         }
 
         public static IWebHostBuilder CreateWebHostBuilder(string[] args) =>
-            WebHost.CreateDefaultBuilder(args)
+              WebHost.CreateDefaultBuilder(args)
+                .ConfigureLogging(c =>
+                {
+                    c.AddFacadeNLog("NLog.config");
+                })
+                .UseFacadeNLog()
                 .UseStartup<Startup>();
     }
 }
